requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
rank-bm25
numpy
tqdm
cohere
voyageai
scikit-learn
colbert-ai
